Disable drop during setup too

+++ tests/global-setup.js
@@ -14,7 +14,14 @@ async function createTestDatabase() {
   });
 
   try {
-    await knex.raw(`DROP DATABASE IF EXISTS ${process.env.DB_NAME}`);
+    console.log("Checking CI:", process.env.CI);
+
+    if (process.env.CI) {
+      console.log("No need to drop db, as container would be destroyed");
+    } else {
+      await knex.raw(`DROP DATABASE IF EXISTS ${process.env.DB_NAME}`);
+      console.log("Test database dropped successfully");
+    }
     await knex.raw(`CREATE DATABASE ${process.env.DB_NAME}`);
   } catch (error) {
     throw new Error(error);
